+++ src/Column.php
@@ -215,7 +215,7 @@ class Column
     /**
      * Wrap the column value into a <a></a> component.
      * You can declare the link as a string or as a closure which will let you manipulate the following attributes :
-     * $model, $column.
+     * \Illuminate\Database\Eloquent\Model $model, \Okipa\LaravelTable\Column $column.
      * If no url is declared, it will be set with the column value.
      *
      * @param string|Closure|null $url
@@ -231,7 +231,8 @@ class Column
 
     /**
      * Display a custom value for the column.
-     * The closure let you manipulate the following attributes : $model, $column.
+     * The closure let you manipulate the following attributes : \Illuminate\Database\Eloquent\Model $model,
+     * \Okipa\LaravelTable\Column$column.
      *
      * @param Closure $valueClosure
      *
@@ -246,7 +247,8 @@ class Column
 
     /**
      * Display a custom HTML for the column.
-     * The closure let you manipulate the following attributes : $model, $column.
+     * The closure let you manipulate the following attributes : \Illuminate\Database\Eloquent\Model $model,
+     * \Okipa\LaravelTable\Column $column.
      *
      * @param Closure $htmlClosure
      *