+++ src/Traits/TableClassesCustomizations.php
@@ -119,7 +119,7 @@ trait TableClassesCustomizations
 
     /**
      * Set rows classes when the given conditions are respected.
-     * The closure let you manipulate the following attribute : $model.
+     * The closure let you manipulate the following attribute : \Illuminate\Database\Eloquent\Model $model.
      *
      * @param \Closure $rowClassesClosure
      * @param array    $rowClasses